Make the getTotalOccurrences methods of the Entry and Postings interfaces return a long, since it's turning out that longs are showing up in some applications.  Make sure that when combining these totals at run time, we don't exceed Long.MAX_VALUE.

+++ src/com/sun/labs/minion/classification/FeaturePostings.java
@@ -115,7 +115,7 @@ public class FeaturePostings implements Postings {
      * Gets the total number of occurrences for these postings, which is
      * just the number of features encoded.
      */
-    public int getTotalOccurrences() {
+    public long getTotalOccurrences() {
         return nIDs;
     }
